feat: Extract messages to filename relative to package.json
affects: babel-plugin-lingui-extract-messages

+++ packages/babel-plugin-lingui-extract-messages/test/index.js
@@ -1,6 +1,6 @@
 import fs from 'fs'
 import path from "path"
-import { transformFileSync } from 'babel-core'
+import {transformFileSync} from 'babel-core'
 
 import plugin from '../src/index'
 
@@ -8,23 +8,24 @@ import plugin from '../src/index'
 const LOCALE_DIR = './locale'
 
 const rmdir = (dir) => {
-    const list = fs.readdirSync(dir)
+  if (!fs.existsSync(dir)) return
+  const list = fs.readdirSync(dir)
 
-    for(let i = 0; i < list.length; i++) {
-        const filename = path.join(dir, list[i])
-        const stat = fs.statSync(filename)
+  for (let i = 0; i < list.length; i++) {
+    const filename = path.join(dir, list[i])
+    const stat = fs.statSync(filename)
 
-        if(filename == "." || filename == "..") {
-            // pass these files
-        } else if(stat.isDirectory()) {
-            // rmdir recursively
-            rmdir(filename);
-        } else {
-            // rm fiilename
-            fs.unlinkSync(filename)
-        }
+    if (filename == "." || filename == "..") {
+      // pass these files
+    } else if (stat.isDirectory()) {
+      // rmdir recursively
+      rmdir(filename);
+    } else {
+      // rm fiilename
+      fs.unlinkSync(filename)
     }
-    fs.rmdirSync(dir)
+  }
+  fs.rmdirSync(dir)
 }
 
 
@@ -44,9 +45,29 @@ function testCase(testName, assertion) {
 }
 
 
-describe('babel-plugin-lingui-extract-messages', function() {
+describe('babel-plugin-lingui-extract-messages', function () {
+  const buildDir = path.join(LOCALE_DIR, '_build', 'test', 'fixtures')
+
+  beforeAll(() => {
+    rmdir(LOCALE_DIR)
+  })
+
   afterAll(() => {
     rmdir(LOCALE_DIR)
+  })
+
+  testCase('should raise exception on duplicate id and different defaults', (transform) => {
+    expect(transform('duplicate-id.js')).toThrow(/Different defaults/)
+  })
+
+  testCase("shouldn't write catalog for files without translatable messages", (transform) => {
+    expect(transform('empty.js')).not.toThrow()
+    expect(fs.existsSync(path.join(buildDir, 'empty.json'))).toBeFalsy()
+  })
+
+  testCase("shouldn't path to file inside locale dir", (transform) => {
+    expect(transform('deep/all.js')).not.toThrow()
+    expect(fs.existsSync(path.join(buildDir, 'deep/all.json'))).toBeTruthy()
   })
 
   testCase('should extract all messages', (transform) => {
@@ -55,29 +76,25 @@ describe('babel-plugin-lingui-extract-messages', function() {
     // another runs should write messages
     expect(transform('all.js')).not.toThrow()
 
-    const messages = JSON.parse(fs.readFileSync(path.join(LOCALE_DIR, '_build/all.json')))
+    const messages = JSON.parse(fs.readFileSync(path.join(buildDir, 'all.json')))
     expect(messages).toEqual({
       "msg.hello": {
         "origin": [
-          ["../test/fixtures/all.js", 2]
+          ["test/fixtures/all.js", 2]
         ]
       },
       "msg.default": {
         "defaults": "Hello World",
         "origin": [
-          ["../test/fixtures/all.js", 3],
-          ["../test/fixtures/all.js", 4]
+          ["test/fixtures/all.js", 3],
+          ["test/fixtures/all.js", 4]
         ]
       },
       "Hi, my name is <0>{name}</0>": {
         "origin": [
-          ["../test/fixtures/all.js", 5]
+          ["test/fixtures/all.js", 5]
         ]
       }
     })
-  })
-
-  testCase('should raise exception on duplicate id and different defaults', (transform) => {
-    expect(transform('duplicate-id.js')).toThrow(/Different defaults/)
   })
 })
